refactor (again)
